apps/tag/src/tag.config.service.spec.ts
import { ConfigService } from "@nestjs/config";
import { TagConfigService } from "./tag.config.service";
import { Test } from "@nestjs/testing";

describe("TagConfigService", () => {
  // TODO: interface ?
  const configServiceMock = {
    get: (_propertyPtagh: string, defaultValue: any): any => defaultValue,
  };

  let service: TagConfigService;

  beforeEach(async () => {
    const module = await Test.createTestingModule({
      providers: [
        TagConfigService,
        {
          provide: ConfigService,
          useValue: configServiceMock,
        },
      ],
    }).compile();
    service = module.get<TagConfigService>(TagConfigService);
  });

  it("should be defined", () => {
    expect(service).toBeDefined();
  });

  it("servicePort should be equal to a value", () => {
    expect(service.servicePort).toEqual(0);
  });

  it("serviceRetryTagtempts should be equal to a value", () => {
    expect(service.serviceRetryAttempts).toEqual(0);
  });

  it("serviceRetryDelay should be equal to a value", () => {
    expect(service.serviceRetryDelay).toEqual(0);
  });

  it("serviceUrl should be equal to a value", () => {
    expect(service.serviceUrl).toEqual("");
  });

  it("typeormDtagabase should be equal to a value", () => {
    expect(service.typeormDatabase).toEqual("");
  });

  it("typeormHost should be equal to a value", () => {
    expect(service.typeormHost).toEqual("");
  });

  it("typeormLogging should be equal to a value", () => {
    expect(service.typeormLogging).toEqual(true);
  });

  it("typeormPassword should be equal to a value", () => {
    expect(service.typeormPassword).toEqual("");
  });

  it("typeormPort should be equal to a value", () => {
    expect(service.typeormPort).toEqual(0);
  });

  it("typeormSynchronize should be equal to a value", () => {
    expect(service.typeormSynchronize).toEqual(true);
  });

  it("typeormUsername should be equal to a value", () => {
    expect(service.typeormUsername).toEqual("");
  });
});
